requirements.txt
fastapi
uvicorn
streamlit
openai
faiss-cpu
opencv-python
pillow
numpy
openai-whisper
python-multipart
